Выстраданное домашнее задание №15. С внесением некоторых изменений и некоторых рекомендаций ревьюера
