+++ app/build.gradle.kts
@@ -42,6 +42,8 @@ android {
 
 dependencies {
 
+    implementation(libs.androidx.core.ktx.v190)
+
     implementation(libs.retrofit)
     implementation(libs.converter.gson)
