--- app/src/main/java/com/practicum/playlistmaker/creator/CreatorAudioPlayer.kt
@@ -1,21 +1,0 @@
-package com.practicum.playlistmaker.creator
-
-import android.media.MediaPlayer
-import com.practicum.playlistmaker.data.audioPlayer.AudioPlayerRepositoryImpl
-import com.practicum.playlistmaker.domain.api.repository.AudioPlayerRepository
-import com.practicum.playlistmaker.domain.impl.AudioPlayerInteractorImpl
-
-object CreatorAudioPlayer {
-
-    private fun getMediaPlayer(): MediaPlayer {
-        return MediaPlayer()
-    }
-
-    private fun getAudioPlayerRepository(): AudioPlayerRepository {
-        return AudioPlayerRepositoryImpl(getMediaPlayer())
-    }
-
-    fun provideAudioPlayerInteractor(): AudioPlayerInteractorImpl {
-        return AudioPlayerInteractorImpl(getAudioPlayerRepository())
-    }
-}
