+++ app/src/main/java/com/practicum/playlistmaker/data/NetworkConst.kt
@@ -1,0 +1,9 @@
+package com.practicum.playlistmaker.data
+
+object NetworkConst {
+
+    const val DARK_THEME_KEY = "dark_theme_mode"
+    const val SHARED_PREFERENCES = "shared_preferences"
+    const val KEY_FOR_HISTORY_TRACK_LIST = "history_track_list"
+
+}
