+++ app/src/main/java/com/practicum/playlistmaker/data/themeSettings/SettingsRepositoryImpl.kt
@@ -6,28 +6,23 @@ import android.content.SharedPreferences
 import android.content.res.Configuration
 import com.practicum.playlistmaker.domain.api.repository.SettingsRepository
 import androidx.core.content.edit
+import com.practicum.playlistmaker.data.NetworkConst
 
 class SettingsRepositoryImpl(private val application: Application) : SettingsRepository {
 
     private val sharedPreferences: SharedPreferences = application.getSharedPreferences(
-        SHARED_PREFERENCES, Context.MODE_PRIVATE
+        NetworkConst.SHARED_PREFERENCES, Context.MODE_PRIVATE
     )
 
     override fun getTheme(): Boolean {
         val darkModeOn = application.resources.configuration.uiMode and
                 Configuration.UI_MODE_NIGHT_MASK == Configuration.UI_MODE_NIGHT_YES
-        return sharedPreferences.getBoolean(DARK_THEME_KEY, darkModeOn)
+        return sharedPreferences.getBoolean(NetworkConst.DARK_THEME_KEY, darkModeOn)
     }
 
     override fun savedTheme(isDarkModeOn: Boolean) {
         sharedPreferences.edit() {
-            putBoolean(DARK_THEME_KEY, isDarkModeOn)
+            putBoolean(NetworkConst.DARK_THEME_KEY, isDarkModeOn)
         }
     }
-
-    companion object {
-        private const val DARK_THEME_KEY = "dark_theme_mode"
-        private const val SHARED_PREFERENCES = "shared_preferences"
-    }
-
 }
