+++ gradle/libs.versions.toml
@@ -2,6 +2,7 @@
 agp = "8.9.2"
 compiler = "4.16.0"
 constraintlayoutVersion = "2.2.1"
+coreKtxVersion = "1.9.0"
 gson = "2.12.1"
 kotlin = "2.0.21"
 coreKtx = "1.15.0"
@@ -19,6 +20,7 @@ uiTextAndroid = "1.8.0"
 [libraries]
 androidx-constraintlayout-v200 = { module = "androidx.constraintlayout:constraintlayout", version.ref = "constraintlayoutVersion" }
 androidx-core-ktx = { group = "androidx.core", name = "core-ktx", version.ref = "coreKtx" }
+androidx-core-ktx-v190 = { module = "androidx.core:core-ktx", version.ref = "coreKtxVersion" }
 compiler = { module = "com.github.bumptech.glide:compiler", version.ref = "compiler" }
 converter-gson = { module = "com.squareup.retrofit2:converter-gson", version.ref = "retrofit" }
 glide = { module = "com.github.bumptech.glide:glide", version.ref = "compiler" }
